+++ NEWS.md
@@ -16,6 +16,7 @@
   [#27](https://github.com/Sbozzolo/kuibit/pull/27))
 - Added fourth order accuracy in finite differences of grid functions
 - Added `copy` to gravitational and electromagnetic waves
+- Added `crop` and `cropped` to multipole data
 
 #### New `tree` module